requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# AI/ML dependencies (optional - remove if not using local models)
transformers==4.35.2
torch>=2.2.0

# Development dependencies (remove in production if needed)
# uvicorn[standard] includes watchfiles, websockets, etc.

# Additional production dependencies
gunicorn==21.2.0  # Alternative WSGI server
starlette==0.27.0  # FastAPI dependency (explicit version)

# Logging and monitoring
structlog==23.2.0  # Better structured logging